requirements.txt
cohere
python-multipart
fastapi
pydantic
uvicorn
python-dotenv
pinecone-client